+++ homeassistant/components/intent/__init__.py
@@ -2,13 +2,8 @@
 
 from __future__ import annotations
 
-import asyncio
-from collections import defaultdict
-from collections.abc import Callable, Coroutine
-from dataclasses import dataclass
-from enum import StrEnum
+from collections.abc import Callable
 import logging
-import time
 from typing import Any, Protocol
 
 from aiohttp import web
@@ -43,9 +38,16 @@ from homeassistant.const import (
 from homeassistant.core import DOMAIN as HA_DOMAIN, HomeAssistant, State
 from homeassistant.helpers import config_validation as cv, integration_platform, intent
 from homeassistant.helpers.typing import ConfigType
-from homeassistant.util import ulid
 
 from .const import DOMAIN, TIMER_DATA
+from .timers import (
+    CancelTimerIntentHandler,
+    DecreaseTimerIntentHandler,
+    IncreaseTimerIntentHandler,
+    SetTimerIntentHandler,
+    TimerHandler,
+    TimerManager,
+)
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -333,451 +335,15 @@ class SetPositionIntentHandler(intent.DynamicServiceIntentHandler):
         raise intent.IntentHandleError(f"Domain not supported: {state.domain}")
 
 
-class TimerEventType(StrEnum):
-    """Timer event type."""
-
-    STARTED = "started"
-    FINISHED = "finished"
-    CANCELLED = "cancelled"
-    UPDATED = "updated"
-
-
-@dataclass(frozen=True)
-class TimerEvent:
-    """Event sent when a timer changes state."""
-
-    type: TimerEventType
-    timer_id: str
-    seconds_left: int
-    name: str | None = None
-
-
-@dataclass
-class TimerInfo:
-    """Information for a single timer."""
-
-    id: str
-    name: str | None
-    seconds: int
-    device_id: str
-    task: asyncio.Task
-    start_hours: int | None
-    start_minutes: int | None
-    start_seconds: int | None
-
-    updated_at: int
-    """Timestamp when timer was last updated (set with time.monotonic_ns)"""
-
-    @property
-    def seconds_left(self) -> int:
-        """Return number of seconds left on the timer."""
-        now = time.monotonic_ns()
-        seconds_running = int((now - self.updated_at) / 1e9)
-        return max(0, self.seconds - seconds_running)
-
-
-TimerHandler = Callable[[TimerEvent], Coroutine[Any, Any, None]]
-
-
-class TimerManager:
-    """Manager for intent timers."""
-
-    def __init__(self, hass: HomeAssistant) -> None:
-        """Initialize timer manager."""
-        self.hass = hass
-
-        # timer id -> timer
-        self.timers: dict[str, TimerInfo] = {}
-
-        # device id -> handlers
-        self.handlers: dict[str, list[TimerHandler]] = defaultdict(list)
-
-    def register_handler(
-        self, device_id: str, handler: TimerHandler
-    ) -> Callable[[], None]:
-        """Register a timer event handler for a device."""
-        self.handlers[device_id].append(handler)
-
-        return lambda: self.handlers[device_id].remove(handler)
-
-    async def start_timer(
-        self,
-        device_id: str,
-        hours: int | None,
-        minutes: int | None,
-        seconds: int | None,
-        name: str | None = None,
-    ) -> str:
-        """Start a timer."""
-        total_seconds = 0
-        if hours is not None:
-            total_seconds += 60 * 60 * hours
-
-        if minutes is not None:
-            total_seconds += 60 * minutes
-
-        if seconds is not None:
-            total_seconds += seconds
-
-        timer_id = ulid.ulid_now()
-        created_at = time.monotonic_ns()
-        self.timers[timer_id] = TimerInfo(
-            id=timer_id,
-            name=name,
-            start_hours=hours,
-            start_minutes=minutes,
-            start_seconds=seconds,
-            seconds=total_seconds,
-            device_id=device_id,
-            task=self.hass.async_create_background_task(
-                self._wait_for_timer(timer_id, total_seconds, created_at),
-                name=f"Timer {timer_id}",
-            ),
-            updated_at=created_at,
-        )
-
-        event = TimerEvent(
-            TimerEventType.STARTED, timer_id, name=name, seconds_left=total_seconds
-        )
-        await asyncio.gather(*(handler(event) for handler in self.handlers[device_id]))
-
-        return timer_id
-
-    async def _wait_for_timer(
-        self, timer_id: str, seconds: int, updated_at: int
-    ) -> None:
-        """Sleep until timer is up. Timer is only finished if it hasn't been updated."""
-        try:
-            await asyncio.sleep(seconds)
-            if (timer := self.timers.get(timer_id)) and (
-                timer.updated_at == updated_at
-            ):
-                await self._timer_finished(timer_id, timer.device_id)
-        except asyncio.CancelledError:
-            pass  # expected when timer is updated
-
-    def find_timer_by_name(self, name: str) -> str | None:
-        """Find a timer by name."""
-        for timer in self.timers.values():
-            if timer.name == name:
-                return timer.id
-
-        return None
-
-    def find_timer_by_start(
-        self, hours: int | None, minutes: int | None, seconds: int | None
-    ) -> str | None:
-        """Find a timer by its starting time."""
-        for timer in self.timers.values():
-            if (
-                (timer.start_hours == hours)
-                and (timer.start_minutes == minutes)
-                and (timer.start_seconds == seconds)
-            ):
-                return timer.id
-
-        return None
-
-    async def cancel_timer(self, timer_id: str) -> None:
-        """Cancel a timer."""
-        if timer := self.timers.pop(timer_id, None):
-            timer.seconds = 0
-            timer.updated_at = time.monotonic_ns()
-            timer.task.cancel()
-            event = TimerEvent(
-                TimerEventType.CANCELLED, timer_id, name=timer.name, seconds_left=0
-            )
-            await asyncio.gather(
-                *(handler(event) for handler in self.handlers[timer.device_id])
-            )
-
-    async def add_time(self, timer_id: str, seconds: int) -> None:
-        """Add time to a timer."""
-        if timer := self.timers.get(timer_id):
-            timer.seconds = max(0, timer.seconds_left + seconds)
-            timer.updated_at = time.monotonic_ns()
-            timer.task.cancel()
-            timer.task = self.hass.async_create_background_task(
-                self._wait_for_timer(timer_id, timer.seconds, timer.updated_at),
-                name=f"Timer {timer_id}",
-            )
-            event = TimerEvent(
-                TimerEventType.UPDATED,
-                timer_id,
-                name=timer.name,
-                seconds_left=timer.seconds,
-            )
-            await asyncio.gather(
-                *(handler(event) for handler in self.handlers[timer.device_id])
-            )
-
-    async def remove_time(self, timer_id: str, seconds: int) -> None:
-        """Remove time from a timer."""
-        await self.add_time(timer_id, -seconds)
-
-    async def _timer_finished(self, timer_id: str, device_id: str) -> None:
-        """Call event handlers when a timer finishes."""
-        if timer := self.timers.pop(timer_id):
-            event = TimerEvent(
-                TimerEventType.FINISHED, timer_id, name=timer.name, seconds_left=0
-            )
-            await asyncio.gather(
-                *(handler(event) for handler in self.handlers[device_id])
-            )
-
-
 def async_register_timer_handler(
-    hass: HomeAssistant, device_id: str, handler: TimerHandler
+    hass: HomeAssistant, handler: TimerHandler, device_id: str | None
 ) -> Callable[[], None]:
     """Register a handler for timer events on a device.
 
     Returns a function to unregister.
     """
     timer_manager: TimerManager = hass.data[TIMER_DATA]
-    return timer_manager.register_handler(device_id, handler)
-
-
-class SetTimerIntentHandler(intent.IntentHandler):
-    """Intent handler for starting a new timer."""
-
-    intent_type = intent.INTENT_SET_TIMER
-    slot_schema = {
-        vol.Required("device_id"): cv.string,
-        vol.Required(vol.Any("hours", "minutes", "seconds")): cv.positive_int,
-        vol.Optional("name"): cv.string,
-    }
-
-    async def async_handle(self, intent_obj: intent.Intent) -> intent.IntentResponse:
-        """Handle the intent."""
-        hass = intent_obj.hass
-        timer_manager: TimerManager = hass.data[TIMER_DATA]
-        slots = self.async_validate_slots(intent_obj.slots)
-
-        device_id: str = slots["device_id"]["value"]
-        name: str | None = None
-        if "name" in slots:
-            name = slots["name"]["value"]
-
-        hours: int | None = None
-        if "hours" in slots:
-            hours = int(slots["hours"]["value"])
-
-        minutes: int | None = None
-        if "minutes" in slots:
-            minutes = int(slots["minutes"]["value"])
-
-        seconds: int | None = None
-        if "seconds" in slots:
-            seconds = int(slots["seconds"]["value"])
-
-        await timer_manager.start_timer(device_id, hours, minutes, seconds, name=name)
-
-        return intent_obj.create_response()
-
-
-class CancelTimerIntentHandler(intent.IntentHandler):
-    """Intent handler for cancelling running timer."""
-
-    intent_type = intent.INTENT_CANCEL_TIMER
-    slot_schema = {
-        vol.Any("start_hours", "start_minutes", "start_seconds"): cv.positive_int,
-        vol.Optional("name"): cv.string,
-    }
-
-    async def async_handle(self, intent_obj: intent.Intent) -> intent.IntentResponse:
-        """Handle the intent."""
-        hass = intent_obj.hass
-        timer_manager: TimerManager = hass.data[TIMER_DATA]
-        slots = self.async_validate_slots(intent_obj.slots)
-
-        response = intent_obj.create_response()
-
-        timer_id: str | None = None
-        if "name" in slots:
-            name: str = slots["name"]["value"]
-            timer_id = timer_manager.find_timer_by_name(name)
-
-            if timer_id is not None:
-                # Matched by name
-                await timer_manager.cancel_timer(timer_id)
-                return response
-
-            response.async_set_error(
-                intent.IntentResponseErrorCode.NO_VALID_TARGETS,
-                f"No timer named {name}",
-            )
-            return response
-
-        start_hours: int | None = None
-        if "start_hours" in slots:
-            start_hours = int(slots["start_hours"]["value"])
-
-        start_minutes: int | None = None
-        if "start_minutes" in slots:
-            start_minutes = int(slots["start_minutes"]["value"])
-
-        start_seconds: int | None = None
-        if "start_seconds" in slots:
-            start_seconds = int(slots["start_seconds"]["value"])
-
-        timer_id = timer_manager.find_timer_by_start(
-            start_hours, start_minutes, start_seconds
-        )
-        if timer_id is not None:
-            # Matched by starting time
-            await timer_manager.cancel_timer(timer_id)
-            return response
-
-        response.async_set_error(
-            intent.IntentResponseErrorCode.NO_VALID_TARGETS,
-            f"No timer for hours={start_hours}, minutes={start_minutes}, seconds={start_seconds}",
-        )
-
-        return response
-
-
-class IncreaseTimerIntentHandler(intent.IntentHandler):
-    """Intent handler for increasing the time of a running timer."""
-
-    intent_type = intent.INTENT_INCREASE_TIMER
-    slot_schema = {
-        vol.Any("hours", "minutes", "seconds"): cv.positive_int,
-        vol.Any("start_hours", "start_minutes", "start_seconds"): cv.positive_int,
-        vol.Optional("name"): cv.string,
-    }
-
-    async def async_handle(self, intent_obj: intent.Intent) -> intent.IntentResponse:
-        """Handle the intent."""
-        hass = intent_obj.hass
-        timer_manager: TimerManager = hass.data[TIMER_DATA]
-        slots = self.async_validate_slots(intent_obj.slots)
-
-        response = intent_obj.create_response()
-
-        total_seconds = 0
-        if "hours" in slots:
-            total_seconds += 60 * 60 * int(slots["hours"]["value"])
-
-        if "minutes" in slots:
-            total_seconds += 60 * int(slots["minutes"]["value"])
-
-        if "seconds" in slots:
-            total_seconds += int(slots["seconds"]["value"])
-
-        timer_id: str | None = None
-        if "name" in slots:
-            name: str = slots["name"]["value"]
-            timer_id = timer_manager.find_timer_by_name(name)
-
-            if timer_id is not None:
-                # Matched by name
-                await timer_manager.add_time(timer_id, total_seconds)
-                return response
-
-            response.async_set_error(
-                intent.IntentResponseErrorCode.NO_VALID_TARGETS,
-                f"No timer named {name}",
-            )
-            return response
-
-        start_hours: int | None = None
-        if "start_hours" in slots:
-            start_hours = int(slots["start_hours"]["value"])
-
-        start_minutes: int | None = None
-        if "start_minutes" in slots:
-            start_minutes = int(slots["start_minutes"]["value"])
-
-        start_seconds: int | None = None
-        if "start_seconds" in slots:
-            start_seconds = int(slots["start_seconds"]["value"])
-
-        timer_id = timer_manager.find_timer_by_start(
-            start_hours, start_minutes, start_seconds
-        )
-        if timer_id is not None:
-            # Matched by starting time
-            await timer_manager.add_time(timer_id, total_seconds)
-            return response
-
-        response.async_set_error(
-            intent.IntentResponseErrorCode.NO_VALID_TARGETS,
-            f"No timer for hours={start_hours}, minutes={start_minutes}, seconds={start_seconds}",
-        )
-
-        return response
-
-
-class DecreaseTimerIntentHandler(intent.IntentHandler):
-    """Intent handler for decreasing the time of a running timer."""
-
-    intent_type = intent.INTENT_DECREASE_TIMER
-    slot_schema = {
-        vol.Any("hours", "minutes", "seconds"): cv.positive_int,
-        vol.Any("start_hours", "start_minutes", "start_seconds"): cv.positive_int,
-        vol.Optional("name"): cv.string,
-    }
-
-    async def async_handle(self, intent_obj: intent.Intent) -> intent.IntentResponse:
-        """Handle the intent."""
-        hass = intent_obj.hass
-        timer_manager: TimerManager = hass.data[TIMER_DATA]
-        slots = self.async_validate_slots(intent_obj.slots)
-
-        response = intent_obj.create_response()
-
-        total_seconds = 0
-        if "hours" in slots:
-            total_seconds += 60 * 60 * int(slots["hours"]["value"])
-
-        if "minutes" in slots:
-            total_seconds += 60 * int(slots["minutes"]["value"])
-
-        if "seconds" in slots:
-            total_seconds += int(slots["seconds"]["value"])
-
-        timer_id: str | None = None
-        if "name" in slots:
-            name: str = slots["name"]["value"]
-            timer_id = timer_manager.find_timer_by_name(name)
-
-            if timer_id is not None:
-                # Matched by name
-                await timer_manager.remove_time(timer_id, total_seconds)
-                return response
-
-            response.async_set_error(
-                intent.IntentResponseErrorCode.NO_VALID_TARGETS,
-                f"No timer named {name}",
-            )
-            return response
-
-        start_hours: int | None = None
-        if "start_hours" in slots:
-            start_hours = int(slots["start_hours"]["value"])
-
-        start_minutes: int | None = None
-        if "start_minutes" in slots:
-            start_minutes = int(slots["start_minutes"]["value"])
-
-        start_seconds: int | None = None
-        if "start_seconds" in slots:
-            start_seconds = int(slots["start_seconds"]["value"])
-
-        timer_id = timer_manager.find_timer_by_start(
-            start_hours, start_minutes, start_seconds
-        )
-        if timer_id is not None:
-            # Matched by starting time
-            await timer_manager.remove_time(timer_id, total_seconds)
-            return response
-
-        response.async_set_error(
-            intent.IntentResponseErrorCode.NO_VALID_TARGETS,
-            f"No timer for hours={start_hours}, minutes={start_minutes}, seconds={start_seconds}",
-        )
-
-        return response
+    return timer_manager.register_handler(handler, device_id=device_id)
 
 
 async def _async_process_intent(
